+++ integrations/clickup/main.py
@@ -1,39 +1,13 @@
-from enum import StrEnum
 from typing import Any, Dict
 from loguru import logger
 from port_ocean.context.ocean import ocean
 from port_ocean.core.ocean_types import ASYNC_GENERATOR_RESYNC_TYPE
-from client import ClickupClient
-
-EVENT_ACTIONS = {
-    "listCreated": "register",
-    "listUpdated": "register",
-    "listDeleted": "unregister",
-    "taskCreated": "register",
-    "taskUpdated": "register",
-    "taskDeleted": "unregister",
-    "spaceCreated": "register",
-    "spaceUpdated": "register",
-    "spaceDeleted": "unregister",
-}
-
-WEBHOOK_TEAM_MAP = {}
-
-
-class ObjectKind(StrEnum):
-    TEAM = "team"
-    SPACE = "space"
-    PROJECT = "project"
-    ISSUE = "issue"
-
-
-async def init_client() -> ClickupClient:
-    client = ClickupClient(
-        ocean.integration_config["clickup_base_url"],
-        ocean.integration_config["clickup_personal_token"],
-        ocean.integration_config["include_archived_projects"],
-    )
-    return client
+from webhook import (
+    handle_webhook_creation,
+    process_webhook_request,
+    init_client,
+    ObjectKind,
+)
 
 
 async def setup_application() -> None:
@@ -44,22 +18,7 @@ async def setup_application() -> None:
             "Without setting up the webhook, you will have to manually initiate resync to get updates from ClickUp."
         )
         return
-    clickup_client = await init_client()
-    async for teams in clickup_client.get_clickup_teams():
-        for team in teams:
-            webhooks = await clickup_client.get_clickup_webhooks(team["id"])
-            for webhook in webhooks:
-                if webhook["endpoint"] == f"{app_host}/integration/webhook":
-                    logger.info(f"Webhook already exists for team {team['id']}")
-                    WEBHOOK_TEAM_MAP[webhook["id"]] = team["id"]
-                    break
-                else:
-                    logger.info(f"Creating webhook for team {team['id']}")
-                    webhook = await clickup_client.create_clickup_webhook(
-                        team["id"], app_host
-                    )
-                    WEBHOOK_TEAM_MAP[webhook["id"]] = team["id"]
-    logger.warning(f"Webhook team map: {WEBHOOK_TEAM_MAP}")
+    await handle_webhook_creation(app_host)
 
 
 @ocean.on_resync(ObjectKind.TEAM)
@@ -86,79 +45,19 @@ async def on_resync_projects(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
         yield projects
 
 
-@ocean.on_resync(ObjectKind.ISSUE)
-async def on_resync_issues(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
+@ocean.on_resync(ObjectKind.TASK)
+async def on_resync_tasks(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
     clickup_client = await init_client()
-    async for issues in clickup_client.get_paginated_issues():
-        logger.info(f"Received batch of {len(issues)} issues")
-        yield issues
-
-
-async def handle_register(
-    clickup_client: Any,
-    entity_id: str,
-    kind: ObjectKind,
-    event_type: str,
-    webhook_id: str,
-) -> None:
-    team_id = WEBHOOK_TEAM_MAP.get(webhook_id)
-    if kind == ObjectKind.ISSUE:
-        entity = await clickup_client.get_single_issue(entity_id)
-    elif kind == ObjectKind.PROJECT:
-        entity = await clickup_client.get_single_project(entity_id, team_id)
-    else:
-        entity = await clickup_client.get_single_space(entity_id, team_id)
-
-    if entity:
-        await ocean.register_raw(kind, [entity])
-        logger.info(f"Registered {kind} for event {event_type}")
-    else:
-        logger.error(f"Handler returned None for entity_id {entity_id}")
-
-
-async def handle_unregister(entity_id: str, kind: ObjectKind, event_type: str) -> None:
-    try:
-        await ocean.unregister_raw(kind, [{"id": entity_id}])
-        logger.info(f"Unregistered {kind} for event {event_type}")
-    except Exception as e:
-        logger.error(f"Exception {e} occurred while attempting to unregister raw")
+    async for tasks in clickup_client.get_paginated_tasks():
+        logger.info(f"Received batch of {len(tasks)} tasks")
+        yield tasks
 
 
 @ocean.router.post("/webhook")
 async def handle_webhook_request(data: Dict[str, Any]) -> Dict[str, Any]:
-    """
-    Handle the webhook request from ClickUp.
-    Events are mapped to the appropriate actions in event_handlers.
-    """
-    clickup_client = await init_client()
-    logger.info(f"Received webhook event of type: {data.get('event')}")
-
-    event_handlers = {
-        "task": ObjectKind.ISSUE,
-        "list": ObjectKind.PROJECT,
-        "space": ObjectKind.SPACE,
-    }
-
-    event_type = data["event"]
-    webhook_id = data["webhook_id"]
-    action = "unregister" if "Deleted" in event_type else "register"
-
-    for key, kind in event_handlers.items():
-        if key in event_type:
-            entity_id = data.get(f"{key}_id")
-            if not entity_id:
-                logger.error(f"No {key}_id found in data for event {event_type}")
-                continue
-            if action == "register":
-                await handle_register(
-                    clickup_client, entity_id, kind, event_type, webhook_id
-                )
-            else:
-                await handle_unregister(entity_id, kind, event_type)
-            break
-
-    logger.info("Webhook event processed")
-    return {"ok": True}
+    logger.info(f"Received webhook event of type: {data}")
+    response = await process_webhook_request(data)
+    return response
 
 
 @ocean.on_start()